+++ packages/ui/src/components/atoms/Label/Label.stories.tsx
@@ -1,0 +1,9 @@
+import { Label } from '@components/atoms/Label/Label';
+
+export default {
+  component: Label,
+};
+
+export const Default = {
+  render: () => <Label>Hello World</Label>,
+};