+++ packages/ui/src/utils/ctw.ts
@@ -1,0 +1,4 @@
+import { ClassValue, clsx } from 'clsx';
+import { twMerge } from 'tailwind-merge';
+
+export const ctw = (...classNames: Array<ClassValue>) => twMerge(clsx(classNames));