+++ services/workflows-service/src/metrics/repository/sql/select-active-users.sql.ts
@@ -1,0 +1,6 @@
+export const selectActiveUsersQuery = `
+select
+id, "firstName", "lastName", "lastActiveAt"
+from "User"
+  where "User"."lastActiveAt"  >= $1
+order by "lastActiveAt" desc`;