+++ services/workflows-service/src/metrics/repository/types/aggregate-average-resolution-time.ts
@@ -1,0 +1,3 @@
+export interface IAggregateAverageResolutionTime {
+  average_time: string | null;
+}